# docs/openapi.yaml
openapi: 3.0.3
info:
  title: 房屋责任盘营销闭环系统 API
  version: 1.0.0
  description: |
    一期接口契约（v1）。统一响应包裹 { code, data, message }，code=0 表示成功。
    鉴权：Authorization: Bearer <jwt>。所有路径以 /api/v1 为前缀。
servers:
  - url: /api/v1
    description: 主服务
...
paths:
...
  /rules:
    get:
      tags: [rule]
      summary: 规则列表（按优先级排序）
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
        - name: status
          in: query
          schema: { type: string, enum: [enabled, disabled] }
        - name: packageId
          in: query
          schema: { type: integer, format: int64 }
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - properties:
                      data:
                        $ref: '#/components/schemas/PagedRule'
    post:
      tags: [rule]
      summary: 新建规则
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RuleUpsert'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - properties:
                      data:
                        $ref: '#/components/schemas/Rule'
  /rules/reorder:
    patch:
      tags: [rule]
      summary: 批量调整优先级
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                items:
                  type: array
                  items:
                    type: object
                    properties:
                      id: { type: integer, format: int64 }
                      priority: { type: integer }
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /rules/preview:
    post:
      tags: [rule]
      summary: 命中预览（抽样同步 / 全量异步）
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RulePreviewRequest'
      responses:
        '200':
          description: 抽样预览结果；若 mode=full 则返回 previewTaskId
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - properties:
                      data:
                        $ref: '#/components/schemas/RulePreviewResult'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 200, default: 20 }
  schemas:
    ApiResponse:
      type: object
      properties:
        code: { type: integer, example: 0 }
        message: { type: string, example: "ok" }
        data: { nullable: true }
      required: [code, message]
...
    RuleCondition:
      type: object
      required: [tag, op]
      properties:
        tag: { type: string, description: 标签键（对应 tag_dictionary.tag_key） }
        op: { type: string, enum: [eq, neq, in, not_in, gte, lte, between, exists] }
        value:
          description: 依据 op 为标量、数组或区间
          nullable: true
    RuleConditionGroup:
      type: object
      properties:
        logic: { type: string, enum: [AND, OR], default: AND }
        conditions:
          type: array
          items:
            $ref: '#/components/schemas/RuleCondition'
    Rule:
      type: object
      properties:
        id: { type: integer, format: int64 }
        name: { type: string }
        packageId: { type: integer, format: int64 }
        priority: { type: integer, description: 数字越小优先级越高 }
        mutexGroupId: { type: integer, format: int64, nullable: true }
        conditionLogic: { type: string, enum: [AND, OR] }
        conditionJson: { $ref: '#/components/schemas/RuleConditionGroup' }
        status: { type: string, enum: [enabled, disabled] }
        remark: { type: string }
    RuleUpsert:
      type: object
      required: [name, packageId, priority, conditionJson]
      properties:
        name: { type: string, maxLength: 128 }
        packageId: { type: integer, format: int64 }
        priority: { type: integer }
        mutexGroupId: { type: integer, format: int64, nullable: true }
        conditionLogic: { type: string, enum: [AND, OR] }
        conditionJson: { $ref: '#/components/schemas/RuleConditionGroup' }
        status: { type: string, enum: [enabled, disabled] }
        remark: { type: string }
    PagedRule:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Rule'
        total: { type: integer }
        page: { type: integer }
        limit: { type: integer }
        hasMore: { type: boolean }
...
    RulePreviewRequest:
      type: object
      properties:
        draft:
          $ref: '#/components/schemas/RuleUpsert'
        mode: { type: string, enum: [sample, full], default: sample }
        sampleSize: { type: integer, default: 50000 }
        scope:
          type: object
          description: 可选抽样范围
          properties:
            region: { type: string }
            city: { type: string }
    RulePreviewResult:
      type: object
      properties:
        previewTaskId: { type: string, nullable: true }
        matchedCount: { type: integer }
        suppressedByMutexCount: { type: integer }
        samples:
          type: array
          items:
            type: object
            properties:
              houseId: { type: string }
              communityName: { type: string }
              matchedRuleId: { type: integer, format: int64 }
              packageId: { type: integer, format: int64 }
        conflicts:
          type: array
          description: 与更高优先级规则的冲突提示
          items:
            type: object
            properties:
              ruleId: { type: integer, format: int64 }
              ruleName: { type: string }
              overlapCount: { type: integer }
...
security:
  - bearerAuth: []